papeletas de jueces, ordenar musica por carpetas para campeonato, coach cards con nombres reserva y ordenadas por orden de slaida

+++ puntuaciones_rutina_code.php
@@ -1,20 +1,18 @@
 <?php
-
+error_reporting(E_ALL);
+ini_set('display_errors', 1);
+    error_reporting(E_ALL);
 include('security.php');
 extract($_POST, EXTR_PREFIX_SAME, "wddx");
 //Añadir registro
 if(isset($_POST['save_btn'])){
-	$licencia = $_POST['licencia'];
-	$apellidos = $_POST['apellidos'];
-	$nombre = $_POST['nombre'];
-	$fecha_nacimiento = $_POST['fecha_nacimiento'];
 	$id_club = $_POST['id_club'];
 
-//	$query="INSERT INTO puntuaciones_jueces (id_panel_juez, id_rutina, id_elemento, nota) VALUES ('".$id_panel_juez1."','".$id_rutina."','".$id_elemento."','".$notaE1J1."')";
     $query = "DELETE FROM puntuaciones_jueces WHERE id_rutina = $id_rutina";
     $query_run = mysqli_query($connection,$query);
 
-	$query="INSERT INTO puntuaciones_jueces (id_panel_juez, id_rutina, id_elemento, nota) VALUES ('".$id_panel_juez1."','".$id_rutina."','1','".$notaE1J1."'),('".$id_panel_juez2."','".$id_rutina."','1','".$notaE1J2."'),('".$id_panel_juez3."','".$id_rutina."','1','".$notaE1J3."'),
+	$query="INSERT INTO puntuaciones_jueces (id_panel_juez, id_rutina, id_elemento, nota) VALUES
+	('".$id_panel_juez1."','".$id_rutina."','1','".$notaE1J1."'),('".$id_panel_juez2."','".$id_rutina."','1','".$notaE1J2."'),('".$id_panel_juez3."','".$id_rutina."','1','".$notaE1J3."'),
     ('".$id_panel_juez4."','".$id_rutina."','1','".$notaE1J4."'),
     ('".$id_panel_juez5."','".$id_rutina."','1','".$notaE1J5."');";
     $query_run = mysqli_query($connection,$query);
@@ -30,7 +28,8 @@ if(isset($_POST['save_btn'])){
     $query = "DELETE FROM puntuaciones_jueces WHERE tipo_ia is Null and nota=0.00 and id_elemento>0 and id_rutina=$id_rutina";
     //mysqli_query($connection,$query);
 
-    $query="INSERT INTO puntuaciones_jueces (id_panel_juez, id_rutina, id_elemento, nota) VALUES ('".$id_panel_juez1."','".$id_rutina."','2','".$notaE2J1."'),('".$id_panel_juez2."','".$id_rutina."','2','".$notaE2J2."'),('".$id_panel_juez3."','".$id_rutina."','2','".$notaE2J3."'),
+    $query="INSERT INTO puntuaciones_jueces (id_panel_juez, id_rutina, id_elemento, nota) VALUES
+	('".$id_panel_juez1."','".$id_rutina."','2','".$notaE2J1."'),('".$id_panel_juez2."','".$id_rutina."','2','".$notaE2J2."'),('".$id_panel_juez3."','".$id_rutina."','2','".$notaE2J3."'),
     ('".$id_panel_juez4."','".$id_rutina."','2','".$notaE2J4."'),
     ('".$id_panel_juez5."','".$id_rutina."','2','".$notaE2J5."');";
     $query_run = mysqli_query($connection,$query);
@@ -44,7 +43,8 @@ if(isset($_POST['save_btn'])){
     $query = "DELETE FROM puntuaciones_jueces WHERE tipo_ia is Null and nota=0.00 and id_elemento>0 and id_rutina=$id_rutina";
     //mysqli_query($connection,$query);
 
-    $query="INSERT INTO puntuaciones_jueces (id_panel_juez, id_rutina, id_elemento, nota) VALUES ('".$id_panel_juez1."','".$id_rutina."','3','".$notaE3J1."'),('".$id_panel_juez2."','".$id_rutina."','3','".$notaE3J2."'),('".$id_panel_juez3."','".$id_rutina."','3','".$notaE3J3."'),
+    $query="INSERT INTO puntuaciones_jueces (id_panel_juez, id_rutina, id_elemento, nota) VALUES
+	('".$id_panel_juez1."','".$id_rutina."','3','".$notaE3J1."'),('".$id_panel_juez2."','".$id_rutina."','3','".$notaE3J2."'),('".$id_panel_juez3."','".$id_rutina."','3','".$notaE3J3."'),
     ('".$id_panel_juez4."','".$id_rutina."','3','".$notaE3J4."'),
     ('".$id_panel_juez5."','".$id_rutina."','3','".$notaE3J5."');";
     $query_run = mysqli_query($connection,$query);
@@ -57,8 +57,9 @@ if(isset($_POST['save_btn'])){
     mysqli_query($connection,$query);
     $query = "DELETE FROM puntuaciones_jueces WHERE tipo_ia is Null and nota=0.00 and id_elemento>0 and id_rutina=$id_rutina";
    // mysqli_query($connection,$query);
-
-    $query="INSERT INTO puntuaciones_jueces (id_panel_juez, id_rutina, id_elemento, nota) VALUES ('".$id_panel_juez1."','".$id_rutina."','4','".$notaE4J1."'),('".$id_panel_juez2."','".$id_rutina."','4','".$notaE4J2."'),('".$id_panel_juez3."','".$id_rutina."','4','".$notaE4J3."'),
+    if($notaE4J1 != ''){
+    $query="INSERT INTO puntuaciones_jueces (id_panel_juez, id_rutina, id_elemento, nota) VALUES
+	('".$id_panel_juez1."','".$id_rutina."','4','".$notaE4J1."'),('".$id_panel_juez2."','".$id_rutina."','4','".$notaE4J2."'),('".$id_panel_juez3."','".$id_rutina."','4','".$notaE4J3."'),
     ('".$id_panel_juez4."','".$id_rutina."','4','".$notaE4J4."'),
     ('".$id_panel_juez5."','".$id_rutina."','4','".$notaE4J5."');";
     $query_run = mysqli_query($connection,$query);
@@ -71,23 +72,27 @@ if(isset($_POST['save_btn'])){
     mysqli_query($connection,$query);
     $query = "DELETE FROM puntuaciones_jueces WHERE tipo_ia is Null and nota=0.00 and id_elemento>0 and id_rutina=$id_rutina";
     //mysqli_query($connection,$query);
-
-    $query="INSERT INTO puntuaciones_jueces (id_panel_juez, id_rutina, id_elemento, nota) VALUES ('".$id_panel_juez1."','".$id_rutina."','5','".$notaE5J1."'),('".$id_panel_juez2."','".$id_rutina."','5','".$notaE5J2."'),('".$id_panel_juez3."','".$id_rutina."','5','".$notaE5J3."'),
-    ('".$id_panel_juez4."','".$id_rutina."','5','".$notaE5J4."'),
-    ('".$id_panel_juez5."','".$id_rutina."','5','".$notaE5J5."');";
-    $query_run = mysqli_query($connection,$query);
-    $query = "SELECT min(nota), max(nota) FROM puntuaciones_jueces WHERE id_rutina=$id_rutina and id_elemento=5";
-    $min_nota = mysqli_fetch_assoc(mysqli_query($connection,$query))['min(nota)'];
-    $max_nota = mysqli_fetch_assoc(mysqli_query($connection,$query))['max(nota)'];
-    $query = "UPDATE puntuaciones_jueces SET nota_menor = 'si' WHERE nota=$min_nota and id_rutina=$id_rutina and id_elemento=5 limit 1";
-    mysqli_query($connection,$query);
-    $query = "UPDATE puntuaciones_jueces SET nota_mayor = 'si' WHERE nota=$max_nota and id_rutina=$id_rutina and id_elemento=5 limit 1";
-    mysqli_query($connection,$query);
-    $query = "DELETE FROM puntuaciones_jueces WHERE tipo_ia is Null and nota=0.00 and id_elemento>0 and id_rutina=$id_rutina";
-    //mysqli_query($connection,$query);
+	}
+    if(isset($notaE5J1)){
+		$query="INSERT INTO puntuaciones_jueces (id_panel_juez, id_rutina, id_elemento, nota) VALUES
+		('".$id_panel_juez1."','".$id_rutina."','5','".$notaE5J1."'),('".$id_panel_juez2."','".$id_rutina."','5','".$notaE5J2."'),('".$id_panel_juez3."','".$id_rutina."','5','".$notaE5J3."'),
+		('".$id_panel_juez4."','".$id_rutina."','5','".$notaE5J4."'),
+		('".$id_panel_juez5."','".$id_rutina."','5','".$notaE5J5."');";
+		$query_run = mysqli_query($connection,$query);
+		$query = "SELECT min(nota), max(nota) FROM puntuaciones_jueces WHERE id_rutina=$id_rutina and id_elemento=5";
+		$min_nota = mysqli_fetch_assoc(mysqli_query($connection,$query))['min(nota)'];
+		$max_nota = mysqli_fetch_assoc(mysqli_query($connection,$query))['max(nota)'];
+		$query = "UPDATE puntuaciones_jueces SET nota_menor = 'si' WHERE nota=$min_nota and id_rutina=$id_rutina and id_elemento=5 limit 1";
+		mysqli_query($connection,$query);
+		$query = "UPDATE puntuaciones_jueces SET nota_mayor = 'si' WHERE nota=$max_nota and id_rutina=$id_rutina and id_elemento=5 limit 1";
+		mysqli_query($connection,$query);
+		$query = "DELETE FROM puntuaciones_jueces WHERE tipo_ia is Null and nota=0.00 and id_elemento>0 and id_rutina=$id_rutina";
+		//mysqli_query($connection,$query);
+	}
     /////////////////////arreglar lo de arriba con un bucle
-    if($notaE6J1 != ''){
-        $query="INSERT INTO puntuaciones_jueces (id_panel_juez, id_rutina, id_elemento, nota) VALUES ('".$id_panel_juez1."','".$id_rutina."','6','".$notaE6J1."'),('".$id_panel_juez2."','".$id_rutina."','6','".$notaE6J2."'),('".$id_panel_juez3."','".$id_rutina."','6','".$notaE6J3."'),
+    if(isset($notaE6J1)){
+        $query="INSERT INTO puntuaciones_jueces (id_panel_juez, id_rutina, id_elemento, nota) VALUES
+		('".$id_panel_juez1."','".$id_rutina."','6','".$notaE6J1."'),('".$id_panel_juez2."','".$id_rutina."','6','".$notaE6J2."'),('".$id_panel_juez3."','".$id_rutina."','6','".$notaE6J3."'),
         ('".$id_panel_juez4."','".$id_rutina."','6','".$notaE6J4."'),
         ('".$id_panel_juez5."','".$id_rutina."','6','".$notaE6J5."');";
         $query_run = mysqli_query($connection,$query);
@@ -101,7 +106,7 @@ if(isset($_POST['save_btn'])){
         $query = "DELETE FROM puntuaciones_jueces WHERE tipo_ia is Null and nota=0.00 and id_elemento>0 and id_rutina=$id_rutina";
     }
     //mysqli_query($connection,$query);
-    if($notaE7J1 != ''){
+    if(isset($notaE7J1)){
         $query="INSERT INTO puntuaciones_jueces (id_panel_juez, id_rutina, id_elemento, nota) VALUES ('".$id_panel_juez1."','".$id_rutina."','7','".$notaE7J1."'),('".$id_panel_juez2."','".$id_rutina."','7','".$notaE7J2."'),('".$id_panel_juez3."','".$id_rutina."','7','".$notaE7J3."'),
         ('".$id_panel_juez4."','".$id_rutina."','7','".$notaE7J4."'),
         ('".$id_panel_juez5."','".$id_rutina."','7','".$notaE7J5."');";
@@ -116,8 +121,9 @@ if(isset($_POST['save_btn'])){
         $query = "DELETE FROM puntuaciones_jueces WHERE tipo_ia is Null and nota=0.00 and id_elemento>0 and id_rutina=$id_rutina";
         //mysqli_query($connection,$query);
     }
-     if($notaE8J1 != ''){
-        $query="INSERT INTO puntuaciones_jueces (id_panel_juez, id_rutina, id_elemento, nota) VALUES ('".$id_panel_juez1."','".$id_rutina."','8','".$notaE8J1."'),('".$id_panel_juez2."','".$id_rutina."','8','".$notaE8J2."'),('".$id_panel_juez3."','".$id_rutina."','8','".$notaE8J3."'),
+    if(isset($notaE8J1)){
+        $query="INSERT INTO puntuaciones_jueces (id_panel_juez, id_rutina, id_elemento, nota) VALUES
+		('".$id_panel_juez1."','".$id_rutina."','8','".$notaE8J1."'),('".$id_panel_juez2."','".$id_rutina."','8','".$notaE8J2."'),('".$id_panel_juez3."','".$id_rutina."','8','".$notaE8J3."'),
         ('".$id_panel_juez4."','".$id_rutina."','8','".$notaE8J4."'),
         ('".$id_panel_juez5."','".$id_rutina."','8','".$notaE8J5."');";
         $query_run = mysqli_query($connection,$query);
@@ -132,7 +138,7 @@ if(isset($_POST['save_btn'])){
         //mysqli_query($connection,$query);
     }
 
-    if($notaE9J1 != ''){
+    if(isset($notaE9J1)){
         $query="INSERT INTO puntuaciones_jueces (id_panel_juez, id_rutina, id_elemento, nota) VALUES ('".$id_panel_juez1."','".$id_rutina."','9','".$notaE9J1."'),('".$id_panel_juez2."','".$id_rutina."','9','".$notaE9J2."'),('".$id_panel_juez3."','".$id_rutina."','9','".$notaE9J3."'),
         ('".$id_panel_juez4."','".$id_rutina."','9','".$notaE9J4."'),
         ('".$id_panel_juez5."','".$id_rutina."','9','".$notaE9J5."');";
@@ -151,63 +157,63 @@ if(isset($_POST['save_btn'])){
     /////////////////////
     $query = "DELETE FROM puntuaciones_elementos WHERE id_rutina = $id_rutina";
     $query_run = mysqli_query($connection,$query);
-    if($BM1 != ''){
+    if(isset($BB1)){
             $dd1=$BM1;
             $llevado_BM = "llevado_BM='si',";
     }else
         $llevado_BM = "";
     $query="INSERT puntuaciones_elementos SET $llevado_BM nota_media = (SELECT (sum(nota)-min(nota)-max(nota))/(count(nota)-2) FROM puntuaciones_jueces WHERE id_rutina=$id_rutina and id_elemento=1), nota=(SELECT (sum(nota)-min(nota)-max(nota))/(count(nota)-2) FROM puntuaciones_jueces WHERE id_rutina=$id_rutina and id_elemento=1)*$dd1*$factor1, id_rutina=$id_rutina, elemento=1";
     mysqli_query($connection,$query);
-    if($BM2 != ''){
+    if(isset($BB2)){
             $dd2=$BM2;
             $llevado_BM = "llevado_BM='si',";
-}else
+	}else
         $llevado_BM = "";
     $query="INSERT puntuaciones_elementos SET $llevado_BM nota_media = (SELECT (sum(nota)-min(nota)-max(nota))/(count(nota)-2) FROM puntuaciones_jueces WHERE id_rutina=$id_rutina and id_elemento=2), nota=(SELECT (sum(nota)-min(nota)-max(nota))/(count(nota)-2) FROM puntuaciones_jueces WHERE id_rutina=$id_rutina and id_elemento=2)*$dd2*$factor2, id_rutina=$id_rutina, elemento=2";
     mysqli_query($connection,$query);
-    if($BM3 != ''){
+    if(isset($BB3)){
             $dd3=$BM3;
             $llevado_BM = "llevado_BM='si',";
     }else
         $llevado_BM = "";
     $query="INSERT puntuaciones_elementos SET $llevado_BM nota_media = (SELECT (sum(nota)-min(nota)-max(nota))/(count(nota)-2) FROM puntuaciones_jueces WHERE id_rutina=$id_rutina and id_elemento=3), nota=(SELECT (sum(nota)-min(nota)-max(nota))/(count(nota)-2) FROM puntuaciones_jueces WHERE id_rutina=$id_rutina and id_elemento=3)*$dd3*$factor3, id_rutina=$id_rutina, elemento=3";
     mysqli_query($connection,$query);
-    if($BM4 != ''){
+    if(isset($BB4)){
             $dd4=$BM4;
             $llevado_BM = "llevado_BM='si',";
     }else
         $llevado_BM = "";
     $query="INSERT puntuaciones_elementos SET $llevado_BM nota_media = (SELECT (sum(nota)-min(nota)-max(nota))/(count(nota)-2) FROM puntuaciones_jueces WHERE id_rutina=$id_rutina and id_elemento=4), nota=(SELECT (sum(nota)-min(nota)-max(nota))/(count(nota)-2) FROM puntuaciones_jueces WHERE id_rutina=$id_rutina and id_elemento=4)*$dd4*$factor4, id_rutina=$id_rutina, elemento=4";
     mysqli_query($connection,$query);
-    if($BM5 != ''){
+    if(isset($BB5)){
             $dd5=$BM5;
             $llevado_BM = "llevado_BM='si',";
     }else
         $llevado_BM = "";
     $query="INSERT puntuaciones_elementos SET $llevado_BM nota_media = (SELECT (sum(nota)-min(nota)-max(nota))/(count(nota)-2) FROM puntuaciones_jueces WHERE id_rutina=$id_rutina and id_elemento=5), nota=(SELECT (sum(nota)-min(nota)-max(nota))/(count(nota)-2) FROM puntuaciones_jueces WHERE id_rutina=$id_rutina and id_elemento=5)*$dd5*$factor5, id_rutina=$id_rutina, elemento=5";
     mysqli_query($connection,$query);
-    if($BM6 != ''){
+    if(isset($BB6)){
             $dd6=$BM6;
             $llevado_BM = "llevado_BM='si',";
     }else
         $llevado_BM = "";
     $query="INSERT puntuaciones_elementos SET $llevado_BM nota_media = (SELECT (sum(nota)-min(nota)-max(nota))/(count(nota)-2) FROM puntuaciones_jueces WHERE id_rutina=$id_rutina and id_elemento=6), nota=(SELECT (sum(nota)-min(nota)-max(nota))/(count(nota)-2) FROM puntuaciones_jueces WHERE id_rutina=$id_rutina and id_elemento=6)*$dd6*$factor6, id_rutina=$id_rutina, elemento=6";
-    mysqli_query($connection,$query);
-    if($BM7 != ''){
+    mysqli_query($connection, $query);
+    if(isset($BB7)){
         $dd7=$BM7;
         $llevado_BM = "llevado_BM='si',";
     }else
         $llevado_BM = "";
     $query="INSERT puntuaciones_elementos SET $llevado_BM nota_media = (SELECT (sum(nota)-min(nota)-max(nota))/(count(nota)-2) FROM puntuaciones_jueces WHERE id_rutina=$id_rutina and id_elemento=7), nota=(SELECT (sum(nota)-min(nota)-max(nota))/(count(nota)-2) FROM puntuaciones_jueces WHERE id_rutina=$id_rutina and id_elemento=7)*$dd7*$factor7, id_rutina=$id_rutina, elemento=7";
     mysqli_query($connection,$query);
-    if($BM8 != ''){
+    if(isset($BB8)){
         $dd8=$BM8;
         $llevado_BM = "llevado_BM='si',";
     }else
         $llevado_BM = "";
     $query="INSERT puntuaciones_elementos SET $llevado_BM nota_media = (SELECT (sum(nota)-min(nota)-max(nota))/(count(nota)-2) FROM puntuaciones_jueces WHERE id_rutina=$id_rutina and id_elemento=8), nota=(SELECT (sum(nota)-min(nota)-max(nota))/(count(nota)-2) FROM puntuaciones_jueces WHERE id_rutina=$id_rutina and id_elemento=8)*$dd8*$factor8, id_rutina=$id_rutina, elemento=8";
     mysqli_query($connection,$query);
-    if($BM9 != ''){
+    if(isset($BB9)){
         $dd9=$BM9;
         $llevado_BM = "llevado_BM='si',";
     }else
@@ -270,7 +276,7 @@ if(isset($_POST['save_btn'])){
     if($penalizaciones == '')
         $penalizaciones = 0;
     //nota final
-    $query = "SELECT sum(nota) - $penalizaciones FROM puntuaciones_elementos WHERE id_rutina=$id_rutina)";
+    $query = "SELECT sum(nota) - $penalizaciones FROM puntuaciones_elementos WHERE id_rutina=$id_rutina";
     $nota_elementos = mysqli_fetch_assoc(mysqli_query($connection,$query))['sum(nota)'];
     $query = "SELECT nota from errores_sincronizacion WHERE id_rutina=$id_rutina";
     $errores_sincro = mysqli_fetch_assoc(mysqli_query($connection,$query))['nota'];
